requirements.txt
aiohttp==3.9.1
matplotlib==3.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0